move scheme data out of profile_bindlink.cc; convert profile_bindlink.cc to testcase runner
