+++ atomspace/query/animals.atomspace.scm
@@ -1,0 +1,4 @@
+(InheritanceLink (ConceptNode "Frog") (ConceptNode "animal"))
+(InheritanceLink (ConceptNode "Zebra") (ConceptNode "animal"))
+(InheritanceLink (ConceptNode "Deer") (ConceptNode "animal"))
+(InheritanceLink (ConceptNode "Spaceship") (ConceptNode "machine"))
