+++ atomspace/query/animals.query.scm
@@ -1,0 +1,10 @@
+(BindLink 
+    (VariableNode "$var")
+    ;; The pattern to be searched for
+    (InheritanceLink 
+        (VariableNode "$var") 
+        (ConceptNode "animal")
+    )
+    ;; The value to be returned.
+    (VariableNode "$var")
+)
